requirements.txt
Werkzeug==1.0.0
joblib==0.14.1
ffmpeg_python==0.2.0
envparse==0.2.0
scipy==1.4.1
moviepy==1.0.3
pandas==1.0.5
numpy==1.19.0
fastapi==0.61.2
uvicorn==0.12.2
scikit-learn==0.22.1
python-multipart==0.0.5
pydantic==1.7.2